PostsPerPage: split sniff
Split this sniff into two: one in VIP checking for disabling of pagination, one in WP checking for high pagination limit.

The WP version now throws a `warning`. This is reset to an error for use by VIP via the VIP ruleset.
